requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Marshmallow
marshmallow-sqlalchemy
python-dotenv
Flask-Swagger-UI
pytest